Add literal expression
